# plugin/openApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AJRM Marine Display API",
    "version": "0.7.3",
    "description": "Read-only Display projections plus authenticated route, voyage-observation and diagnostic controls. Traffic, Notifications and Audio remain the authoritative providers."
  },
  "servers": [{ "url": "/plugins/signalk-ajrm-marine-display" }],
  "paths": {
    "/status": { "get": { "summary": "Get Display runtime status and browser defaults", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/getTargets": { "get": { "summary": "Get Traffic targets projected for Display", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/getCollisionProfiles": { "get": { "summary": "Get Traffic profile thresholds formatted for Display", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/uiState": { "get": { "summary": "Get the consolidated Display UI projection", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/panelEvents": { "get": { "summary": "Get active alert-panel entries", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/alertEvents": { "get": { "summary": "Get active visual and browser-audio alert entries", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/announcementLog": { "get": { "summary": "Get provider-authored announcement history", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/browserSpeechEvents": { "get": { "summary": "Get the latest explicitly authorized browser-speech delivery", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/autoProfileStatus": { "get": { "summary": "Get Traffic Auto Profile state projected for Display", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/debugControls": {
      "get": { "summary": "Get transient rendering debug controls", "responses": { "200": { "$ref": "#/components/responses/Ok" } } },
      "post": { "summary": "Update transient rendering debug controls", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" } } }
    },
    "/repeatIntervals": { "get": { "summary": "Get Display help repeat-interval overrides", "responses": { "200": { "$ref": "#/components/responses/Ok" } } } },
    "/routes": { "get": { "summary": "List the active route, Signal K route resources and configured Pi GPX files", "responses": { "200": { "$ref": "#/components/responses/Ok" }, "500": { "$ref": "#/components/responses/Error" } } } },
    "/routes/import": { "post": { "summary": "Import a browser GPX route", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/routes/open-pi": { "post": { "summary": "Open a GPX route from the configured Pi directory", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/routes/open-resource": { "post": { "summary": "Open a Signal K route resource", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/routes/delete-resource": { "post": { "summary": "Delete one Signal K route resource", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/routes/reverse": { "post": { "summary": "Reverse the displayed route order", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/routes/save": { "post": { "summary": "Save or save-as the displayed route", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/routes/close": { "post": { "summary": "Close the displayed route", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/routes/export": { "get": { "summary": "Download the active route as GPX 1.1", "responses": { "200": { "description": "GPX route document", "content": { "application/gpx+xml": {} } }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/harbourRegions": { "get": { "summary": "List harbour region resources projected for Display", "responses": { "200": { "$ref": "#/components/responses/Ok" }, "500": { "$ref": "#/components/responses/Error" } } } },
    "/refreshDiagnostics": { "post": { "summary": "Submit an authenticated browser performance diagnostic", "security": [{ "signalk": ["readwrite"] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "500": { "$ref": "#/components/responses/Error" } } } },
    "/observations/status": { "get": { "summary": "Get current Capture voyage-observation capability", "responses": { "200": { "$ref": "#/components/responses/Ok" }, "500": { "$ref": "#/components/responses/Error" } } } },
    "/observations": {
      "post": {
        "summary": "Append a timestamped observation to the active Capture voyage",
        "security": [{ "signalk": ["readwrite"] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["text"], "properties": { "text": { "type": "string", "minLength": 1, "maxLength": 2000 }, "includeSnapshot": { "type": "boolean" } } } } } },
        "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    }
  },
  "components": {
    "securitySchemes": { "signalk": { "type": "http", "scheme": "bearer" } },
    "responses": {
      "Ok": { "description": "Successful response" },
      "Forbidden": { "description": "Signal K read/write or administrator access required" },
      "Error": { "description": "Operation failed" }
    }
  }
}